% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_spc_data}
\alias{example_spc_data}
\title{Example SPC Data}
\format{
\subsection{\code{example_spc_data}}{

A data frame with 1584 rows and 10 columns:
\describe{
\item{indicator}{The unique identifier for the indicator itself}
\item{group}{The unique identifier for the group/establishment}
\item{period_end}{The end date of each SPC period}
\item{period_start}{The start date of each SPC period}
\item{numerator}{The numerator value for the indicator, group and period
combination}
\item{denominator}{The denominator value for the indicator, group and
period combination}
\item{multiplier}{The multiplier applied to the indicator}
\item{better_is}{The direction of favourability for the indicator. I.e. set
to "Lower" when the indicator is improving as it decreases, and set to
"Higher" when the indicator is improving as it increases.}
\item{spc_chart_type}{The type of SPC used in the qiverse.data.qipatterns
package}
\item{funnel_chart_type}{The type of funnel plot used in the FunnelPlotR
package}
}
}
}
\usage{
example_spc_data
}
\description{
Example set of indicators used as an input SPC Funnel Matrix Prep Data Step
}
\keyword{datasets}
